+++ CHANGES.md
@@ -2,4 +2,8 @@
 
 ## 0.1.0
 
-- First release 
+- First release
+- Added the following actions:
+  - doc_lib_list - Returns a list of Document Libraries from the given SP site or subsite
+  - sites_list - Returns a list of all top-levell sharepoint sites at the given URL
+  - subsites_list - Returns a list of all subsites from the given SP site